Clean up server after network sending fails
This fixes the delayed segfault/disconnection aspect of #796.

When sending fails, the server is marked as disconnected and no longer
useful. However it was not yet fully cleaned up. That cleanup happened in
irc_parse_incoming() when we receive traffic from the server. This meant
there could be a noticeable delay until we disconnect. With this change
we clean it up in the next tick.